requirements.txt
requests==2.26.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
python-dotenv==0.19.0
psycopg2-binary==2.9.3
passlib==1.7.4
flask_sqlalchemy==2.5.1
